third_party/WebKit/LayoutTests/platform/win/external/wpt/IndexedDB/idbrequest_result-expected.txt

This is a testharness.js-based test.
PASS IDBRequest.result throws if ready state is pending 
Harness: the test ran to completion.
